Image Identification / Classification through CNN
In this work, a convolutional neural network (CNN) was developed for image classification using the
CIFAR-10 database. The database contains 60,000 color images distributed in 10 classes, 50,000 for training and
10,000 for testing. The network architecture was composed of multiple convolutional layers, followed by pooling and
fully connected layers. The objective is to highlight relevant features of the images. The model was trained using the
backpropagation technique with gradient descent optimization. Results demonstrated the effectiveness of CNN for
the image classification task, with an accuracy of 99% achieved in the test set. The results indicate that the proposed
approach is efficient for classifying images from the CIFAR-10 database, comparable to other techniques in the
literature